requirements.txt
colorama==0.4.6
httpx==0.24.1
pillow==10.0.0